Avoid going into 'Paused' status during long lasting migrations
If a migration is taking longer than 'migration_timeout' the VM
moves into 'Paused' state on the destination host.
This patch increases the timeout to 6 hours as an absolute maximum
for the destination to wait for the migration to finish.
This is now reflected as the migration_destination_timeout.

If the VM after the timeout still has the state PAUSED and the
reason is `migration` we're going to raise an MigrationError and
will lead to the destruction of the VM on the destination.

In all other cases we're keeping the previous behaviour and are
continuing with normal VDSM tasks without touching the state of
the VM. That means that if the VM is in PAUSED state due to other
reasons than migration, it will stay PAUSED.

Change-Id: I6bb1c9ae7ead92093c0d300df7c3567ab20b1e09
Bug-Url: https://bugzilla.redhat.com/1059129
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21963
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23939