netconf: Make vdsm-restore-net-config use _inRollback
The setup networks operations performed by vdsm-restore-net-config
are by nature rollback operations. They are split into two steps,
- removal of the nets.
- configuration of the persisted network config.

Each of the preceding actions has a different rollback context,
making a possible memory rollback (if _inRollback was not used)
rather fruitless, specially when booting up when there is nothing
to roll back to. It is better that the operations are set as
not elegible for rollback.

Change-Id: Iee212bdd6bbf5bcc27dc0a85b2fe6f690ac6dcd0
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/21977
Reviewed-by: Assaf Muller <amuller@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>