QoS: make report only when present and say so in the schema
Libvirt XML format specifies that bandwidth elements (inbound and
outbound) are only optionally present. When present they should
have an average and optionally have the fields 'burst' and 'peak'.

While this was correctly defined in the vdsm api schema for input
(setupNetworks for instance), it was not defined for the capabilities
reporting. That left a not well defined situation where no QoS was
defined and it was being included in the report without following
the specification in the paragraph above.

This patch fixes this inconsistency by reporting the bandwidth
elements only when present in the network definition, and only
containing those fields that are defined for them.

Change-Id: Id9d2bc8310fed2ad3c0a748cdf1cfd2cf050aa2f
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22018
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>