sp: Fix stopping domain monitors
Commit 2671777c69 fixed stopping of domain monitors by stopping monitors from
StoragePool.__del__. This was a dangerous fix because a stray reference to the
pool may delay invocation of __del__ until the referring object is deleted,
delaying stopping of domain monitors for unlimited time. In case of a memory
leak, the pool may never be deleted and domain monitors would never stop.

Fortunately, this fix did seem to work for couple of years, until commit
7b1cc6a20f made the domain monitor pool independent. Instead of one domain
monitor object per pool, there is now a single shared domain monitor object
used by all pool objects. Having a shared domain monitor, the pool __del__
method became deadly.

Shortly after commit 7b1cc6a20f was merged, a new random error appeared in CI
jobs, where there is no storage space left, while storage has plenty of space.
The root cause of the error was anonymous thread running at unexpected times
and stopping silently all domain monitors.  Adding some logging revealed that
this thread was started from StoragePool.__del__ method. This thread was
running 14 minutes after the original pool was disconnected, stopping domain
monitors used by the current pool object. It seems that a task was holding a
reference to the old pool, and when the task was finished, the old pool was
finally destroyed.

We seem to stop monitoring when we should, and stopping monitoring in
the __del__ method is redundant.  This patch removes the deadly method.

Change-Id: Iad80984ab44dd4a4e36b92330eb9320cf68f1580
Relates-To: https://bugzilla.redhat.com/705058
Relates-To: http://gerrit.ovirt.org/19762
Bug-Url: https://bugzilla.redhat.com/1032925
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22058
Reviewed-by: Liron Ar <laravot@redhat.com>
Reviewed-by: Sergey Gotliv <sgotliv@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>