threadPool: Do not keep reference to tasks
Worker thread was taking a task from the threadpool queue, run it, and
then hold it until the next task is run. This causes the task and all
objects referenced from it to keep living until the next task is run by
this thread.

This patch ensure that worker thread does not keep referece to tasks
after they were run.

Change-Id: Ib6e024e0727e11a4c98f4e5ede20229d148fdc58
Relates-To: https://bugzilla.redhat.com/1032925
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22136
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>