hook: extnet: attach a vNIC to a libvirt network defined outside oVirt
This patch adds a hook which allows to connect a vNIC to a libvirt
network that is managed outside of oVirt. To use this hook, you should
first define and activate such a network, as documented in
http://libvirt.org/formatnetwork.html, on each host where the hook is
deployed.

For example, to use an openvswitch bridge, first virsh net-define
    <network>
      <name>ovs-net</name>
      <forward mode='bridge'/>
      <bridge name='ovsbr0'/>
      <virtualport type='openvswitch'/>
    </network>

This preperation was not required in a previous version for the hook
(named forcebridge). However, this version allows to benefit of
non-bridged libvirt networks, too.

Change-Id: I70bc74a9ac4a6eb88d089baa86918275c61f5190
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Humble Chirammal <hchiramm@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22124
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Assaf Muller <amuller@redhat.com>