asyncproc: Removed the close count.
When I originally wrote the class I assumed that since the process side
FD is closed I should mark it closed on VDSMs side as well.

This appeared to be problematic as stuff in the python infrastructure
don't like it when I autoclose the FD for them.

I partially solved it by adding the close counter so that things don't
crash because they don't expect the FD to get closed while they are
reading even if they got EOF.

But since closing the FD doesn't really need to be reflected in the
stream I now decided it's better to just leave it marked as open in the
stream wrapper. The real FD is managed by AsyncProc so there is no worry
about an FD leak.

I just remove the close count buffer and never mark the FD as closed an
leave it up to the user. We loose the ability to detect a stream closing
without waiting for an empty string but that is how python wants you to
do it and this is how we will.

Change-Id: Iaa3bbe8409d0cb0591555e595b3202e9e01c0db4
Signed-off-by: Saggi Mizrahi <smizrahi@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22243
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Antoni Segura Puimedon <asegurap@redhat.com>