netinfo.getRouteDeviceTo: return a device name
As of commit 1a825c79, netinfo.getRouteTo() returned a Route object if
it had found one, and the empty string on various error conditions.
That's a bad practice, and makes BindingXMLRPC code explode with an
AttributeError when it tries to compute route.device.

This patch renames the function to getRouteDeviceTo() and makes it
return routing device name (or empty string) as needed by
BindingXMLRPC.getServerInfo.

Change-Id: I428129cc626c00fd530cec64be2a8c2b9c65e532
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22358
Reviewed-by: Assaf Muller <amuller@redhat.com>