fileSD: Fix image deletion on gluster domain
getImagePath() does not work for glusterSD. The problem is caused by
that domaindir's base paths of glusterSD and other file-based SD are not
the same. For glusterSD, it's sd.storage_repository/mnt/glusterSD,
while it's sd.storage_repository/mnt/ for other file-based storage
domains. This makes the deletion fail on gluster domain.

With this patch we use the storage domain's domaindir to get the image's
directory correctly.

Reference:
  * http://lists.ovirt.org/pipermail/users/2013-December/018653.html

Change-Id: I8ee6cb963a428f18b6e43f0585d461d0975c377a
Signed-off-by: huntxu <mhuntxu@gmail.com>
Bug-Url: https://bugzilla.redhat.com/1046600
Reviewed-on: http://gerrit.ovirt.org/22359
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22563
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>