Rename migration_timeout to migration_downtime_delay
Previously the migration_timeout value was supposed to be used for
multiple things. Now the only meaning left is the calculation for
the delay between the downtime steps.
To reflect this, the value has been renamed and the code updated to make
this meaning more obvious.

Now it is 75 seconds per GiB of RAM the minimum time still is 150 seconds.

Change-Id: I6297b2714c37c98df6af50ed9799df2ef9c596c1
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22421
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>