vdsm: Add support for Guest Reboot
Added optional parameter reboot to shutdown in vdsm and vdsClient.
This parameter represents whether the user wants to reboot the VM
(reboot=True) or just a regular shutdown (reboot=False). To keep
backwards compatibility, the default is False.

Current implementation only utilizes the guest agent for reboot and does
nothing in its absence. The semantics of existing shutdown should be
unchanged by this patch.

Note that older versions of GA don't support the reboot flag and
consqeuently will issue a shutdown instead of reboot.

Change-Id: Ic117574365fc6f7a0dd6e69ba203b2d82d4cefa1
Signed-off-by: Martin Betak <mbetak@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22440
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>