hsm: refresh pool connection on connectStoragePool
When a storage domain is not accessible when connecting to the pool, the
domain links are not created. Later, during the recovery flow, when
engine ask vdsm to reconnect to the pool, the domain links are not
created since the host is already connected to the pool.

This patch ensures that domain structure is created in this case, by
triggering a refresh when the host is already connected to the pool.

Note about poorly documented changes in teh original patch:

- Removing support for auto-reconnection, when msdUUID and masterVersion
  are None. In this case they were read from the pool file by invoking
  getPoolParams(). While this change is not correct, the auto-reconnect
  code is broken for years, (and was removed later in master), so this
  support is not needed.

- Validation of host id is replaced by comparing hostID to pool.id in
  _updateStoragePool().

- pool.getMasterDomain calls was removed becuase it is called in the
  refresh flow later.

Bug-Url: https://bugzilla.redhat.com/1066409
Change-Id: I46d7c36b48f5d58edbc89947ccd2bcd60408a729
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22467
Reviewed-by: Ayal Baron <abaron@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26349
Tested-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>