hsm: refresh pool connection on connectStoragePool
This patch ensures that connectStoragePool would trigger a refresh
when the host is already connected to the same pool.
Using connectStoragePool instead of refreshStoragePool is at the base
of the storage pool metadata removal but it also solves an existing
problem in the non-operational recovery flow where the engine sends
a connectStoragePool request.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1026697
Change-Id: I46d7c36b48f5d58edbc89947ccd2bcd60408a729
Signed-off-by: Federico Simoncelli <fsimonce@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22467
Reviewed-by: Ayal Baron <abaron@redhat.com>