vdsm: Handling topology for ppc64
This change adds a method to extract the CPU topology information
thourght the lscpu command. The current method for x86_64 hasn't
been changed.

This is done due to libvirt showing wrong information about ppc64
topology (issues BZ 621546 and 751205), the lscpu command is being
used to obtain it.

The topology tests were changed to be executed in a mock
architecture and new test cases were added to handle ppc64 methods.

Change-Id: I048d4a6c083392d63fbcff76453e682b9d6f03fc
Signed-off-by: Leonardo Bianconi <leonardo.bianconi@eldorado.org.br>
Signed-off-by: Vitor de Lima <vitor.lima@eldorado.org.br>
Reviewed-on: http://gerrit.ovirt.org/19875
Reviewed-by: Dan Kenigsberg <danken@redhat.com>