vm: Unify checks for vdsm image
When testing if dict or vm.Drive object are vdsm image, we have to do
an ugly type check, and then invoke either vm.Drive.isVdsmImage, or
vm.isVdsmImage, which use different logic.

Having vm.Drive.isVdsmImage looks, at first glance, as an improvement,
but since we handle both dicts and drive objects, this only complicates
the code. The different logic creates confusion and leads to pointless
discussions if some key is needed or not in certain context.

This patch unifies these checks; both vm.Drive and dict have now similar
interface so isVdsmImage can handle both, and we use the same logic.

Change-Id: Iaa109a19aeec56f09ed2e6d64dee636c7779d426
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22370
Reviewed-by: Michal Skrivanek <michal.skrivanek@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Sergey Gotliv <sgotliv@redhat.com>
Reviewed-by: Allon Mureinik <amureini@redhat.com>
Reviewed-by: Federico Simoncelli <fsimonce@redhat.com>