vmfex: Add vmfex device custom properties hook
The current vmfex hook works at the VM level which creates some
usability issues. This new hook makes it easier for the
administrator to map a vNIC to a VM-FEX port by just making him
specify a custom device property called 'vmfex' that has the
Port Profile as content.

This is half of the implementation of level IV of:
http://www.ovirt.org/Features/UCS_Integration

The other half to follow up in another patch is to filter out the
Virtual Functions from the nics report in netinfo.nics()

Change-Id: I830e37c1f3163e7e80bf3c568ff3a06136c39e13
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22529
Reviewed-by: Dan Kenigsberg <danken@redhat.com>