stats: report detailed VM down status
When a VM goes down it reports an exitCode and an
exitMessage which is passed verbatim to the VM log.

In order to have translatable messages, a more detailed
exit reason code is needed.
This patch is the first step towards an explicit, detailed
exit reason value. It laids the ground for a full resolution.

A new explicit exit reason code field is introduced in the
response, to be filled with the detailed status code.
The field is added separately for backward
compatibility with the existing interface.

Future work will split the GenericError code into detailed
codes.

Change-Id: I8d7064fe79d1cd34499fbb32ed0644757cbe05dd
Bug-Url: https://bugzilla.redhat.com/557125
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22631
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>