setupNetworks: support networks with no underlying NIC
The old addNetwork API supported the ability to create a bridged network
that has no leg to the outer world. This capability was lost in the
not-so-very-new setupNetworks API.

If a user had a nicless network defined by addNetwork, and tried to
upgrade to unified persistence, the upgrade script (which uses
setupNetworks) would have failed.

Change-Id: I2d2342cf3629eb05f4249a4657dc2ba45466ba24
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22595
Reviewed-by: Assaf Muller <amuller@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>