ipwrapper: fix excessive process creation
_detectType is a very very often used operation and it was relying
executing the ethtool binary for its operation. That was so costly
that we were forced to use memoization to minimize the impact.
Unfortunately, that came with its own risks, as the memoization
was name based and could eventually give wrong names or even more
common, fail to detect that a device is no longer present in the
system.

python-ethtool has a method for getting this information but
reports the IOErrors incorrectly (missing errnos). A patch has
been submitted for it and when it is released across the OS
versions we support we could drop this code.

Change-Id: I8de502669dfa00585d8d775f51be5617cd4f63d0
Signed-off-by: Antoni S. Puimedon <asegurap@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22609
Reviewed-by: Dan Kenigsberg <danken@redhat.com>