spice: allow to disable the to clipboard copypaste
since version 0.9.3, libvirt allows to optionally disable copy and paste
in the spice protocolol:

http://libvirt.org/formatdomain.html#elementsGraphics

this patch expose this capability in VDSM, in the VM definition.
The new parameter is optional and enabled by default for backward
compatibility.

No XML changes are produced at all unless the new parameter is
specified.

Change-Id: I15f3d74860b99d805d1d204704ec06afcd729484
Bug-Url: https://bugzilla.redhat.com/1052959
Bug-Url: https://bugzilla.redhat.com/800155
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/26916
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>