vm: janitorial: introduce a recovering attribute
So far, a recovering VM was identified by checking for the
presence of a 'recover' attribute in the Vm.conf dict.
Moreover, the attribute was accessed in a thread-unsafe way.
This added unnecessary complication.

This patch makes the code cleaner and simpler introducing
an explicit attribute to identify recovering VMs.

Full history in the discussion of this change:
http://gerrit.ovirt.org/#/c/22556/

Change-Id: I1a43c5f6427c81a6610d820d9475b375c8dc0947
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22783
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>