tests: do not check permissions on dos filesystem
commit 0b147dc2e247e787ad760ba30a8b8815eb65daca
introduced an extra step in unit tests, to check
the permissions of the files in the created filesystem.

This is fine for filesystems which support permissions,
but DOSFS/FAT/VFAT does not, then the tests on this
filesystem started to fail for no good reasons.

This patch fixes this misbehaviour by disabling the
permissions check in filesystem which do not support
them.

Change-Id: I9c14b3cc0538e2f8b09ea99a185ec86e067217f6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22816
Reviewed-by: Vered Volansky <vvolansk@redhat.com>
Tested-by: Vered Volansky <vvolansk@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>