BindingXMLRPC bugfix: Added netinfo:getDeviceByIP and a test
In essence a revert for http://gerrit.ovirt.org/#/c/21470/
I decided not to actually revert so that we keep the new getRouteTo
and its test, as its covered by a test and may be useful in the
future.

Both 'getDeviceByIP' and 'getRouteTo' solve the same issue:
How to report to the engine the device through which the last client
communicated to VDSM. This is used only in installation, so the
engine knows on what NIC to place the management bridge.

The getDeviceByIP approach gets the IP addr
of VDSM's side of the socket, and looks up the NIC which has that IP.
That had issues with testing (The changes can be seen in this patch).
getRouteTo's new (But flawed) approach was to route to the engine's
side of the socket. This failed in all-in-ones as it returned the
loopback device - It's not possible, nor does it make any sense
to place the management bridge on the loopback device.

Change-Id: I3563ae9b3e199e29ea2227463b9e4a3752cb99a0
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22841
Reviewed-by: Dan Kenigsberg <danken@redhat.com>