vdsm: prepareVolumePath payload misdetection fix
The changeset http://gerrit.ovirt.org/#/c/22324/3
used a condition too broad for its check, causing
a device path misdetection when cloud-init is used.

It is triggered in the following case:
specParams vmPayload attribute has
"file"(with content), path" non existent.
Device's (drive's) "path" is set to ''.
The code then consider the cdrom with vmPayload
a case of empty CDROM and creating it as such.
It did work before because the condition results
false when there is no 'path' in drive['specParams']
(which is the case here) and goes to the
next condition section (the right one with payload handling)

This patch provides a band-aid fix to the issue
by tightening the prepareVolumePath check.

Change-Id: Ie847eb92f28d24f4d8189e03a47cd64cc6597363
Bug-Url: https://bugzilla.redhat.com/1047356
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22925
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>