BindingXMLRPC bugfix: Fix netinfo:getIfaceByIP unit test
testGetIfaceByIP fails on machines that have an updated
ethtool, because the test now iterates on devices
that were previously not iterated over.

The test now ignores IPv6 link-local addresses (Which may
be shared with several network devices), and checks
if the device has an ipv4 address configured before
checking it.

Change-Id: Iee4757958ea5554043c6dc7169a8df05b1887687
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22968
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>