fencing: use deathSignal to keep scripts at bay
Before this patch, the fenceNode verb used a knowingly-raceful code to
wait for pending fencing scripts, and kill them when Vdsm exited.
This patch makes sure that deathSignal is sent to pending scripts if
Vdsm exits or crashes.

We require version 1.3 of python-cpopen that raises a proper exception
when an executable is missing.

Change-Id: I068acf5ca0ad0813aaef660770de79958bd0b763
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/22997
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>