Gluster API verbs will now be callable again
Fix regression introduced in #22837.

When wrapping GlusterAPI methods via wrapApiMethod,
wrapApiMethod was accessing a GlusterAPI method's im_self,
(If f is the wrapped method, then f.im_self), which
is an alias to an GlusterAPI instance, but the glusterAPI class
doesn't define updateTimestamp.

This patch moves 'updateTimestamp' to be a module function,
which will be available no matter what method wrapApiMethod
is wrapping.

Change-Id: Ibab01ef654e7ac469ea0a2594c4744918ae24819
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23044
Reviewed-by: Bala.FA <barumuga@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>