Gluster API verbs will now be callable again
Fix regression introduced in #22967.

When wrapping GlusterAPI methods via wrapApiMethod,
wrapApiMethod was accessing a GlusterAPI method's im_self,
(If f is the wrapped method, then f.im_self), which
is an alias to an GlusterAPI instance, but the glusterAPI class
doesn't define updateTimestamp.

This patch moves 'updateTimestamp' to be a module function,
which will be available no matter what method wrapApiMethod
is wrapping.

Relates-To: https://bugzilla.redhat.com/1048596
Change-Id: I1513a5dca444c3e176fb31a9425da85980401b0f
Signed-off-by: Assaf Muller <amuller@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23045
Tested-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Yaniv Bronhaim <ybronhei@redhat.com>