Moving response implementation of getAllVmStats to API.py
This code was introduced as part of jsonrpc work.

getAllVmStats command was implemented in xml binding whereas it should
be in API. Xml binging uses API directly and json binding uses bridge to
proxy calls to API. Having this logic in API instead of xml binging
makes it available for both protocols.

Change-Id: Ic86f5daf5def88ea4fecad85cf2b1aa5a919cc39
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: http://gerrit.ovirt.org/23376
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>