introducing capability to stream data to image
This patch introduces the a new capabaility to vdsm
which allows to upload using streaming content to vdsm
images.

Previously we sent ovf data using XMLRPC (UpdateVM verb),
which limits the size of the data, having to encode the
payload into the xml, and make it hard and inefficient
to upload lot of data and store it on some image.

This patch adds the capabillity to stream data to image,
allowing efficient upload of data in any size and format
and storing it directly on an
image. As the  XML-RPC spec doesn't support streaming and
to avoid requiring another port by using dedicated http
server, in this patch we use the existing xmlrpc server to
handle upload requests.

General upload information:
-----------------------------------------------------------
PUT requests arriving to the server with content type of
application/octet-stream to default paths that we use
today for request handling ('/', '/RPC2') will be treated
as upload requests.
The upload itself is being executed within a task, that's
needed to indicate that there's an operation executed by
the host.

Change-Id: I768b84799ed9fb2769c6d4240519d036f8988b99
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23281
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>