Removing unnecessary check-local from storage Makefile.am
We used that to run storage_exception main [1] to verify the numeration of
the error code numbers. This section was removed from storage_exception
and is not needed anymore. Therefore this patch omits the redundant
check-local scope.

[1] http://gerrit.ovirt.org/#/c/2356/4/vdsm/storage/storage_exception.py

This fixes a regression introduced in commit ed426e5 which adds import
to zombiereaper in utils.py. With the removal of this check-local we
don't require zombiereaper during build time.

Change-Id: I81c634c3218ca6d21bce57fc9c500a906f8f0775
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23183
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23202