sp: changing un-shared treatment in the createVolume function
Up until now, the only way to create a 'template' disk was by using the
copyImage command.  This has 2 limitations:
1. It forces the user to copy all the data instead of doing an in-place
convert of an image to a template image
2. Resulting image cannot be qcow which is wasteful in some cases

This patch addresses that, and allows using an existing disk as a
template disk, by allowing to "share" a disk that isn't marked as
"shared", but can be shared as it doesn't have snapshots. This disk will
be marked as "shared" upon first use.

Change-Id: Id9a4b84795b176b9696684045e9774d324a42eca
Signed-off-by: Oved Ourfali <oourfali@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23260
Reviewed-by: Ayal Baron <abaron@redhat.com>