spice: allow to disable the agent file transfer
Spice has agent based file transfer support.
This option can be disabled in QEMU, and  libvirt has support
to pass this flag to the underlying QEMU.

This patch expose this capability in VDSM, in the VM definition,
implementing the VDSM part of bz983019.
The new parameter is optional and enabled by default for backward
compatibility.

No XML changes are produced at all unless the new parameter is
specified.

Change-Id: Icc98a346baca4ab163ff30e891ae80535e49caa4
Bug-Url: https://bugzilla.redhat.com/983019
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/26917
Reviewed-by: Dan Kenigsberg <danken@redhat.com>