caps: check availability abort on EIO in libvirt
commit bbeb165e42673cddc87495c3d12c4a7f7572013c
added support for VIR_MIGRATE_ABORT_ON_ERROR,
but libvirt 1.0.5.8 shipped in Fedora 19
do not support that, even if it should be supported
starting from upstream version 1.0.1.

This patch provides a workaround for this; if libvirt
misses this flag, it forces the the cluster level at
no more than version 3.3.

Change-Id: I7dd857ac3c00c48d4de3a50931cb259f45ff4f5d
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23273
Reviewed-by: Dan Kenigsberg <danken@redhat.com>