To allow downgrade of vdsm we need to support vdsm-tool old API
During upgrading or downgrading vdsm version we perform reconfigure
process. This process has been changed between 3.3 to 3.4. This patch
tries the current version implementation of reconfigure, and if it fails
it will fallback to vdsmd reconfigure verb, which exists in all
previous versions of vdsm.

This workaround need to be removed in vdsm 3.5.

Change-Id: I8bd7f9048acc4f53f3da66e21d5ad2c5caf2a5ad
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23333
Reviewed-by: Alon Bar-Lev <alonbl@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Douglas Schilling Landgraf <dougsland@redhat.com>