netinfo: ifcfg: ignore 0 suffix
'IPADDR0', 'GATEWAY0', 'PREFIX0' and 'NETMASK0' means to initscripts
exactly as their unsuffixed siblings. With this patch, Vdsm reports them
as 'IPADDR', so the do not confuse Engine.

A proper change would be to make Engine/Vdsm API independent of
ifcfg-specific details. This patch is required anway, for the sake of
older Engines.

Change-Id: I5c9a206a5b96d0b38aab9c3d62d6074424b419be
Bug-Url: https://bugzilla.redhat.com/987832
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: http://gerrit.ovirt.org/23329
Reviewed-by: Assaf Muller <amuller@redhat.com>
Reviewed-by: Antoni Segura Puimedon <asegurap@redhat.com>
Reviewed-by: Sandro Bonazzola <sbonazzo@redhat.com>