Retry connection: SSLEngine not reusable
When connection is lost and there is reconnect SSLEngine can not be
reused so each new connection needs to create new instance of SSLEngine.

When vdsm fails badly the engine side socketChannel assumes that the
connection is still operational but none of the request can be
successfully sent after vdsm reboot. Mitigation for this situation is to
count in the row failed calls and if there is more then default value
try to reconnect.


Change-Id: I03554224af125e6ee899e8845914fa2d9f9c6bcf
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>